- Cleaned up Gantt printing code a bit.
git-svn-id: https://www.taskjuggler.org/svn/taskjuggler/trunk@1150 e1914e07-63f8-0310-9059-d6d858d7cdca

+++ TaskJugglerUI/TjPrintReport.h
@@ -66,9 +66,9 @@ protected:
     const Report* reportDef;
     const QtReportElement* reportElement;
 
-    int mmToXPixels(int mm);
-    int mmToYPixels(int mm);
-    int pointsToYPixels(int pts);
+    int mmToXPixels(double mm);
+    int mmToYPixels(double mm);
+    int pointsToYPixels(double pts);
 
     int scenario;
     TaskList taskList;
